requirements.txt
flask==2.3.3
redis==5.0.1
rq==1.15.1
gunicorn==21.2.0
torch==2.3.1
faster-whisper==0.10.1
yt-dlp==2024.11.4
transformers==4.44.2
sentencepiece==0.2.0
colorama==0.4.6
stripe==7.9.0
python-dotenv==1.0.0
openai>=1.13.3
